requirements.txt
flask
langchain
chromadb
sentence-transformers
openai
pypdf
langchain-community
python-dotenv
